# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BNBU Sports Student API
  version: 1.0.0
  description: Android student client contract. Successful responses are not wrapped in a data envelope.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
    description: Same-origin API prefix; deployment supplies the HTTPS host.
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in a student
      operationId: studentLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /sport/summary:
    get:
      tags: [Sport]
      summary: Get sport-hour progress
      operationId: sportSummary
      responses:
        "200":
          description: Student sport progress
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SportSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /sport/records:
    get:
      tags: [Sport]
      summary: List sport records
      operationId: listSportRecords
      parameters:
        - { in: query, name: courseId, schema: { type: string } }
        - { $ref: "#/components/parameters/PageLimit" }
        - { $ref: "#/components/parameters/PageOffset" }
      responses:
        "200":
          description: Records owned by the authenticated student
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SportRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Sport]
      summary: Submit a sport record
      operationId: createSportRecord
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSportRecord" }
      responses:
        "201":
          description: Record submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MutationResult" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /sport/records/{id}:
    parameters:
      - { $ref: "#/components/parameters/ResourceId" }
    get:
      tags: [Sport]
      summary: Get a sport record
      operationId: getSportRecord
      responses:
        "200":
          description: Record detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SportRecord" }
        "404": { $ref: "#/components/responses/NotFound" }
  /sport/identity:
    get:
      tags: [Sport]
      summary: List sport identities
      operationId: sportIdentity
      responses:
        "200":
          description: Team and club memberships
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Membership" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /common/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      operationId: listNotifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /common/notifications/{id}/read:
    parameters:
      - { $ref: "#/components/parameters/ResourceId" }
    put:
      tags: [Notifications]
      summary: Mark a notification read
      operationId: markNotificationRead
      responses:
        "200":
          description: Read state
          content:
            application/json:
              schema:
                type: object
                required: [id, read]
                properties:
                  id: { type: string }
                  read: { type: boolean, const: true }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/student/push-devices:
    post:
      tags: [Notifications]
      summary: Register the current Android device for generic FCM alerts
      operationId: registerPushDevice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PushDeviceRegistration" }
      responses:
        "204": { description: Device token registered }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      tags: [Notifications]
      summary: Remove the current Android device from FCM delivery
      operationId: unregisterPushDevice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PushDeviceRegistration" }
      responses:
        "204": { description: Device token removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/student/feedback:
    get:
      tags: [Feedback]
      summary: List the current student's problem-feedback tickets
      operationId: listFeedbackTickets
      responses:
        "200":
          description: Tickets submitted by the authenticated student
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackTicketList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Feedback]
      summary: Submit a problem report
      operationId: createFeedbackTicket
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFeedbackTicket" }
      responses:
        "201":
          description: Problem report accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackTicket" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /student/profile:
    get:
      tags: [Student]
      summary: Get the student profile
      operationId: getStudentProfile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [Student]
      summary: Update the student profile
      operationId: updateStudentProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [gender]
              properties:
                gender: { type: [string, "null"], enum: [male, female, null] }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/student/contacts/email/send-code:
    post:
      tags: [Student]
      summary: Send an email contact-verification code
      operationId: sendEmailContactCode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailContactRequest" }
      responses:
        "204": { description: Code accepted for delivery }
        "422": { $ref: "#/components/responses/ValidationError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /v1/student/contacts/email/verify:
    post:
      tags: [Student]
      summary: Verify and bind an email contact
      operationId: verifyEmailContactCode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyEmailContactRequest" }
      responses:
        "200":
          description: Updated profile; activation may now be complete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/student/contacts/phone/send-code:
    post:
      tags: [Student]
      summary: Send a mobile contact-verification code
      operationId: sendPhoneContactCode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PhoneContactRequest" }
      responses:
        "204": { description: Code accepted for delivery }
        "422": { $ref: "#/components/responses/ValidationError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /v1/student/contacts/phone/verify:
    post:
      tags: [Student]
      summary: Verify and bind a mobile contact
      operationId: verifyPhoneContactCode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyPhoneContactRequest" }
      responses:
        "200":
          description: Updated profile; activation may now be complete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/student/preferences/language:
    put:
      tags: [Student]
      summary: Store the language used for server-originated communication
      operationId: updateStudentLanguagePreference
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LanguagePreference" }
      responses:
        "200":
          description: Stored language preference
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LanguagePreference" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageLimit:
      in: query
      name: limit
      required: false
      description: Number of newest rows to return; defaults to 500.
      schema: { type: integer, minimum: 1, maximum: 500, default: 500 }
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ValidationError:
      description: Invalid request
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    RateLimited:
      description: Too many requests
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [account, password]
      properties:
        account: { type: string }
        password: { type: string, format: password }
        role: { type: string, const: student, default: student }
        clientType: { type: string, const: mobile, default: mobile }
    LoginResponse:
      type: object
      required: [token, user, defaultRoute]
      properties:
        token: { type: string }
        defaultRoute: { type: string }
        user: { $ref: "#/components/schemas/User" }
...
security:
  - bearerAuth: []